Overhaul container image for recent PowerDNS version
This simplifies the image by removing the previous runsv based service
system, replacing it with a single plain PowerDNS process. The newer
versions of PowerDNS already have Prometheus metrics built into the HTTP
API and can therefore provide metrics themselves. Also the restriction
for environment variables to ones listed in an allow list has been
lifted, so any config can be expressed, regardless of the version of the
image.